--- polymorphia-frontend/app/(logged-in)/profile/types.ts
@@ -1,1 +1,0 @@
-export type ProfileFilterId = "rankingOptions";
