+++ polymorphia-frontend/components/user-points/types.ts
@@ -4,6 +4,5 @@ export interface UserPointsProps {
   separators?: boolean;
   titleSize?: Size;
   xpSize?: Size;
-  maxCols?: number;
   xpDetails: Record<string, string>;
 }
