+++ polymorphia-frontend/views/profile/types.ts
@@ -1,0 +1,7 @@
+import { Size } from "@/interfaces/general";
+
+export interface UserPointsProps {
+  className: string;
+  titleSize: Size;
+  xpSize: Size;
+}
